Forgot to add init script cleanup and removed securityContexts from the yaml file as they are no longer supported

+++ kubelet.sh
@@ -1,12 +1,9 @@
 #!/bin/sh
 
-KUBE_CERT=/var/run/kubernetes/kubelet/cert.pem
-KUBE_KEY=/var/run/kubernetes/kubelet/key.pem
-
 run_start() {
 	echo "Starting kubelet..."
 	if [ -f $KUBE_CERT ] && [ -f $KUBE_KEY ]; then 
-		/usr/local/bin/kubelet --api-servers=http://127.0.0.1:8080 --allow-privileged=true --config=/etc/kubernetes/manifests --tls-cert-file=$KUBE_CERT --tls-private-key-file=$KUBE_KEY --v=2 > /var/log/kubelet.log 2>&1 &
+		/usr/local/bin/kubelet --api-servers=http://127.0.0.1:8080 --allow-privileged=true --config=/etc/kubernetes/manifests --kubeconfig=/etc/kubernetes/kubelet.kubeconfig --v=2 > /var/log/kubelet.log 2>&1 &
 	else
 		/usr/local/bin/kubelet --api-servers=http://127.0.0.1:8080 --allow-privileged=true --config=/etc/kubernetes/manifests --v=2 > /var/log/kubelet.log 2>&1 &
 	fi
@@ -15,7 +12,12 @@ run_start() {
 
 run_stop() {
 	echo "Stopping kubelet..."
-	kill -9 $(pidof kubelet)
+	P="x$(pidof kubelet)"
+
+	if [ $P != "x" ]; then
+		kill -9 $(pidof kubelet)
+		# TODO: Add something to restart the containers as well
+	fi
 }
 
 run_restart() {
